requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.0.0
opencv-python>=4.7.0
tensorboard>=2.13.0
